Conflicts:
	omgcms/src/main/java/com/omgcms/web/action/api/ResourceActionController.java

+++ omgcms/src/main/java/com/omgcms/web/action/api/UserController.java
@@ -89,7 +89,7 @@ public class UserController {
 
 	
 	@RequestMapping(value = "/delete/{userId}", method = RequestMethod.DELETE)
-	public boolean deleteUser(@PathVariable(value = "userId") Integer userId) {
+	public boolean deleteUser(@PathVariable(value = "userId") Long userId) {
 
 		userId = ParamUtil.get(userId, -1);
 
@@ -145,7 +145,7 @@ public class UserController {
 		Date now = new Date();
 		user.setCreateDate(now);
 		user.setModifyDate(now);
-
+		
 		User savedUser = userService.saveAndFlush(user);
 
 		return savedUser;
